+++ lib/fit_loading_indicator.dart
@@ -21,7 +21,7 @@ class FitLoadingIndicator extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final ThemeData appTheme = Theme.of(context);
-    final FitTheme? fitTheme = FitTheme.of(context);
+    final FitTheme? fitTheme = FitTheme.maybeOf(context);
 
     return Container(
       alignment: alignment ?? Alignment.center,