+++ lib/flutter_fit_utils_ui.dart
@@ -33,3 +33,4 @@ export 'fit_rule_validator/fit_rule_validator.dart';
 export 'fit_rule_validator/validator.dart';
 export 'fit_dialog.dart';
 export 'fit_menu/fit_menu_theme_data.dart';
+export 'fit_loading_handler/status_loading_indicator.dart';